+++ src/rlv12/setdone.cpp
@@ -1,7 +1,7 @@
 #include "rlv12.h"
 
 // Set done and possibly errors
-void RLV12::setDone (int32_t status)
+void RLV12::setDone (u16 status)
 {
     rlcs |= status | CSR_DONE;
     if (rlcs & CSR_IE)
